finished upgrade to openssl 3.0

+++ entity/c/c_crypto.h
@@ -5,7 +5,7 @@
 
 #include <openssl/rand.h>
 #include <openssl/err.h>
-#include <openssl/aes.h> 
+#include <openssl/aes.h>
 #include <openssl/pem.h>
 #include <openssl/bio.h>
 #include <openssl/ssl.h>
@@ -27,53 +27,51 @@
 #define MAC_KEY_SIZE 32
 #define CIPHER_KEY_SIZE 16
 #define RSA_KEY_SIZE 256
-#define RSA_ENCRYPT_SIGN_SIZE RSA_KEY_SIZE*2
+#define RSA_ENCRYPT_SIGN_SIZE RSA_KEY_SIZE * 2
 
 #define IV_SIZE 16
 
 #define SHA256_DIGEST_LENGTH 32
 
-
 typedef struct
 {
     unsigned char data[RSA_KEY_SIZE];
     unsigned char sign[RSA_KEY_SIZE];
-}signed_data_t;
+} signed_data_t;
 
-//must free mac & cipher key
+// must free mac & cipher key
 typedef struct
 {
-    unsigned char * mac_key;
+    unsigned char *mac_key;
     unsigned int mac_key_size;
-    unsigned char * cipher_key;
+    unsigned char *cipher_key;
     unsigned int cipher_key_size;
     unsigned char absvalidity[DIST_KEY_EXPIRATION_TIME_SIZE];
     long int start_time;
-}distribution_key_t;
+} distribution_key_t;
 
-
-//must free mac & cipher key
+// must free mac & cipher key
 typedef struct
 {
     unsigned char key_id[SESSION_KEY_ID_SIZE];
     unsigned char abs_validity[KEY_EXPIRATION_TIME_SIZE];
     unsigned char rel_validity[KEY_EXPIRATION_TIME_SIZE];
-    unsigned char * mac_key;
+    unsigned char *mac_key;
     unsigned int mac_key_size;
-    unsigned char * cipher_key;
+    unsigned char *cipher_key;
     unsigned int cipher_key_size;
-}session_key_t; 
+} session_key_t;
 
 void print_last_error(char *msg);
-unsigned char * public_encrypt(unsigned char * data, int data_len, int padding, const char * path, unsigned int *ret_len); 
-unsigned char * private_decrypt(unsigned char * enc_data, int enc_data_len, int padding, const char * path, unsigned int *ret_len);
-unsigned char * SHA256_sign(unsigned char *encrypted, unsigned int encrypted_length, const char * path, unsigned int * sig_length);
-void SHA256_verify(unsigned char * data, unsigned int data_length, unsigned char * sign, unsigned int sign_length, const char * path);
-unsigned char * digest_message_SHA_256(unsigned char *message, int message_length, unsigned int *digest_len);
-void AES_CBC_128_encrypt(unsigned char * plaintext, unsigned int plaintext_length, unsigned char * key, unsigned int key_length, unsigned char * iv, unsigned int iv_length, unsigned char * ret, unsigned int * ret_length);
-void AES_CBC_128_decrypt(unsigned char * encrypted, unsigned int encrypted_length, unsigned char * key, unsigned int key_length, unsigned char  * iv, unsigned int iv_length, unsigned char * ret, unsigned int * ret_length);
+unsigned char *public_encrypt(unsigned char *data, size_t data_len, int padding, const char *path, size_t *ret_len);
+unsigned char *private_decrypt(unsigned char *enc_data, size_t enc_data_len, int padding, const char *path, size_t *ret_len);
+unsigned char *SHA256_sign(unsigned char *encrypted, unsigned int encrypted_length, const char *path, size_t *sig_length);
+void SHA256_verify(unsigned char *data, unsigned int data_length, unsigned char *sign, size_t sign_length, const char *path);
+unsigned char *digest_message_SHA_256(unsigned char *message, int message_length, unsigned int *digest_len);
+void AES_CBC_128_encrypt(unsigned char *plaintext, unsigned int plaintext_length, unsigned char *key, unsigned int key_length, unsigned char *iv, unsigned int iv_length, unsigned char *ret, unsigned int *ret_length);
+void AES_CBC_128_decrypt(unsigned char *encrypted, unsigned int encrypted_length, unsigned char *key, unsigned int key_length, unsigned char *iv, unsigned int iv_length, unsigned char *ret, unsigned int *ret_length);
 
-unsigned char * symmetric_encrypt_authenticate(unsigned char * buf, unsigned int buf_length, unsigned char * mac_key, unsigned int mac_key_size, unsigned char * cipher_key, unsigned int cipher_key_size, unsigned int iv_size, unsigned int * ret_length);
-unsigned char * symmetric_decrypt_authenticate(unsigned char * buf, unsigned int buf_length, unsigned char * mac_key, unsigned int mac_key_size, unsigned char * cipher_key, unsigned int cipher_key_size, unsigned int iv_size, unsigned int * ret_length);
+unsigned char *symmetric_encrypt_authenticate(unsigned char *buf, unsigned int buf_length, unsigned char *mac_key, unsigned int mac_key_size, unsigned char *cipher_key, unsigned int cipher_key_size, unsigned int iv_size, unsigned int *ret_length);
+unsigned char *symmetric_decrypt_authenticate(unsigned char *buf, unsigned int buf_length, unsigned char *mac_key, unsigned int mac_key_size, unsigned char *cipher_key, unsigned int cipher_key_size, unsigned int iv_size, unsigned int *ret_length);
 
 #endif // C_CRYPTO_H
